feat(api-helpers): refactor code from loopback-apps into api-helpers
