+++ modules/api-system/common/models/system/system.js
@@ -1,9 +1,11 @@
 'use strict';
 
+const { getAppConfigs } = require('@colmena/api-helpers')
+
 module.exports = function(System) {
 
   System.ping = () => System.app.models.Ping.ping()
 
-  System.modules = () => System.app.models.Core.getModules()
+  System.modules = () => Promise.resolve(getAppConfigs())
 
 }
