+++ modules/api-system/package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "license": "MIT",
   "dependencies": {
+    "@colmena/api-helpers": "0.0.0",
     "@colmena/logger": "0.0.0",
     "config": "^1.26.1",
     "inflection": "^1.12.0",
