+++ packages/api-helpers/src/helpers.js
@@ -1,8 +1,11 @@
-const prequire = require('parent-require')
-const config = require('config')
-const path = require('path')
+const Promise = require('bluebird')
 
+const config = require('config')
 const log = require('@colmena/logger')
+const path = require('path')
+const prequire = require('parent-require')
+
+const parentRequire = moduleName => prequire(moduleName);
 
 const getModuleList = () => config.get('colmena.modules')
 const getActiveModules = () => Object.keys(getModuleList()).filter(module => getModuleList()[module])
@@ -39,16 +42,22 @@ const dbUpdate = db => {
 }
 
 
+const getAppConfigs = () => getModules().map(moduleName => parentRequire(moduleName))
+
+const getAppFileList = prop => getAppConfigs()
+  .filter(app => app[prop])
+  .map(app => app[prop].map(file => `${app.name}/${file}`))
+  .reduce((res, item) => res.concat(item), [])
+
 const loadModulePath = (moduleName, file) => path.join(moduleName, file);
 
-const loadParentModule = moduleName => prequire(moduleName);
 
 const loadModules = config => {
 
   getModules()
     .forEach(moduleName => {
 
-      const module = loadParentModule(moduleName)
+      const module = parentRequire(moduleName)
       log.white.b(`[loopback-modules] Registering models from module: ${moduleName}`)
 
       if (module.models) {
@@ -73,10 +82,44 @@ const loadModules = config => {
 }
 
 
+const getDomain = (app, domain) => app.models.SystemDomain.findById('default')
+
+const importSampleFileSet = (app, set) => Promise.all(
+  Object.keys(set).map(domainName => {
+    const files = set[domainName]
+    log.gray('[sample-files]', `${files.length} files for domain`, domainName)
+
+    return getDomain(app, domainName)
+      .then(domain => files.forEach(file => domain
+        .importFileByUrl(file.url, file.fileName)
+        .then(() => log.gray('[sample-files]', `${domainName} import ${file.url}`))
+      ))
+  })
+)
+
+const importSampleDataSet = (app, set) => Promise.all(Object.keys(set)
+  .map(modelName => app.models[modelName]
+    .create(set[modelName])
+    .then(res => log.gray('[sample-data]', `${res.length} items for model`, modelName))
+  )
+)
+
+const importSampleData = (app) => Promise.resolve(getAppFileList('sampleData'))
+  .then(items => items.map(item => parentRequire(item)))
+  .then(sets => Promise.each(sets, set => importSampleDataSet(app, set)))
+
+const importSampleFiles = (app) => Promise.resolve(getAppFileList('sampleFiles'))
+  .then(items => items.map(item => parentRequire(item)))
+  .then(sets => Promise.each(sets, set => importSampleFileSet(app, set)))
+
 
 module.exports = {
   dbMigrate,
   dbUpdate,
+  getAppConfigs,
+  getAppFileList,
   getModules,
+  importSampleData,
+  importSampleFiles,
   loadModules,
 }
